Frontend/fix persist

+++ backend/routes/employees.js
@@ -1,4 +1,6 @@
 const router = require('express').Router();
+const validator = require('validator');
+const checkError = require('../utils/validator');
 let Employee = require('../models/employee.model');
 
 router.route('/').get((req, res) => {
@@ -8,11 +10,28 @@ router.route('/').get((req, res) => {
 })
 
 router.route('/').post((req, res) => {
-    const name = req.body.name;
-    const birthDate = Date.parse(req.body.birthDate);
-    const gender = req.body.gender;
-    const salary = Number(req.body.salary);
+    let name = req.body.name;
+    let birthDate = (new Date(req.body.birthDate)).toISOString().substring(0, 10);
+    let gender = req.body.gender;
+    let salary = req.body.salary;
+	const today = (new Date()).toISOString().substring(0, 10);
+	let errors = [];
 
+	if (checkError(name, 'string', 'name', 1, 50, / /))
+		errors.push(...checkError(name, 'string', 'name', 1, 50, / /));
+	if (checkError(birthDate, 'date', 'birthDate', '1951-01-01', today))
+		errors.push(...checkError(birthDate, 'date', 'birthDate', '1951-01-01', today));
+	if (checkError(gender, 'string', 'gender', null, null, null, ['m', 'f']))
+		errors.push(...checkError(gender, 'string', 'gender', null, null, null, ['m', 'f']));
+	if (checkError(salary, 'number', 'salary', 0))
+		errors.push(...checkError(salary, 'number', 'salary', 0));
+
+	if (Object.keys(errors).length)
+		return res.status(422).json({ code: 422, errors});
+
+	
+	birthDate = Date.parse(birthDate);
+	salary = Number(salary);
     const newEmployee = new Employee({
         name, birthDate, gender, salary
     });
@@ -22,6 +41,14 @@ router.route('/').post((req, res) => {
         .catch(err => res.status(400).json('Error: ' + err));
 })
 
+router.route('/analytics').get((req, res) => {
+	Employee.aggregate([
+		{ $group: { _id: '$gender', averageSalary: {$avg: '$salary', } } }
+	])
+		.then(data => res.json(data))
+		.catch(err => res.status(400).json('Error: ' + err));
+});
+
 router.route('/:id').get((req, res) => {
     Employee.findById(req.params.id)
         .then(employee => res.json(employee))
@@ -29,13 +56,34 @@ router.route('/:id').get((req, res) => {
 })
 
 router.route('/:id').put((req, res) => {
+	if (!validator.isMongoId(req.params.id))
+		return res.status(422).json({code: 422, error: "Invalid id"});
+
     Employee.findById(req.params.id)
         .then(employee => {
-            employee.name = req.body.name;
-            employee.birthDate = Date.parse(req.body.birthDate);
-            employee.gender = req.body.gender;
-            employee.salary = Number(req.body.salary);
+            let name = req.body.name;
+			let birthDate = (new Date(req.body.birthDate)).toISOString().substring(0, 10);
+			let gender = req.body.gender;
+			let salary = req.body.salary;
+			const today = (new Date()).toISOString().substring(0, 10);
+			let errors = [];
 
+			if (checkError(name, 'string', 'name', 1, 50, / /))
+				errors.push(...checkError(name, 'string', 'name', 1, 50, / /));
+			if (checkError(birthDate, 'date', 'birthDate', '1951-01-01', today))
+				errors.push(...checkError(birthDate, 'date', 'birthDate', '1951-01-01', today));
+			if (checkError(gender, 'string', 'gender', null, null, null, ['m', 'f']))
+				errors.push(...checkError(gender, 'string', 'gender', null, null, null, ['m', 'f']));
+			if (checkError(salary, 'number', 'salary', 0))
+				errors.push(...checkError(salary, 'number', 'salary', 0));
+
+			if (Object.keys(errors).length)
+				return res.status(422).json({ code: 422, errors});
+
+			employee.name = name;
+			employee.birthDate = Date.parse(birthDate);
+			employee.gender = gender;
+			employee.salary = Number(salary);
             employee.save()
                 .then((e) => res.json(e))
                 .catch(err => res.status(400).json('Error: ' + err));
@@ -44,9 +92,12 @@ router.route('/:id').put((req, res) => {
 });
 
 router.route('/:id').delete((req, res) => {
+	if (!validator.isMongoId(req.params.id))
+		return res.status(422).json({code: 422, error: 'Invalid id'});
+
     Employee.findByIdAndDelete(req.params.id)
         .then(() => res.json('Employee deleted'))
         .catch(err => res.status(400).json('Error: ' + err));
-})
+});
 
 module.exports = router;
